+++ cli/azd/cmd/usage_test.go
@@ -1,10 +1,14 @@
 package cmd
 
 import (
+	"bytes"
+	"html/template"
+	"strings"
 	"testing"
 
 	"github.com/azure/azure-dev/cli/azd/test/snapshot"
 	"github.com/spf13/cobra"
+	"github.com/stretchr/testify/require"
 )
 
 // To update snapshots (assuming your current directory is cli/azd):
@@ -22,7 +26,9 @@ func TestUsage(t *testing.T) {
 
 func usageSnapshot(t *testing.T, cmd *cobra.Command) {
 	t.Run(cmd.Name(), func(t *testing.T) {
-		snapshot.SnapshotT(t, cmd.UsageString())
+		result, err := resolveTemplate(cmd.HelpTemplate(), cmd)
+		require.NoError(t, err)
+		snapshot.SnapshotT(t, result)
 
 		for _, c := range cmd.Commands() {
 			if !c.IsAvailableCommand() || c.IsAdditionalHelpTopicCommand() {
@@ -32,4 +38,16 @@ func usageSnapshot(t *testing.T, cmd *cobra.Command) {
 			usageSnapshot(t, c)
 		}
 	})
+}
+
+func resolveTemplate(text string, data interface{}) (string, error) {
+	finalBuffer := &bytes.Buffer{}
+	t := template.New("resolve template with command")
+	template.Must(t.Parse(text))
+
+	if err := t.Execute(finalBuffer, data); err != nil {
+		return "", err
+	}
+	// update `>` and `<`
+	return strings.ReplaceAll(strings.ReplaceAll(finalBuffer.String(), "&lt;", "<"), "&gt;", ">"), nil
 }