+++ cli/azd/pkg/project/service_target_containerapp_test.go
@@ -4,12 +4,14 @@
 package project
 
 import (
+	"context"
 	"fmt"
 	"strings"
 	"testing"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	"github.com/azure/azure-dev/cli/azd/pkg/infra"
+	"github.com/azure/azure-dev/cli/azd/test/mocks"
 	"github.com/benbjohnson/clock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -18,52 +20,45 @@ import (
 func TestNewContainerAppTargetTypeValidation(t *testing.T) {
 	t.Parallel()
 
-	t.Run("ValidateTypeSuccess", func(t *testing.T) {
-		_, err := NewContainerAppTarget(
-			nil,
-			nil,
-			environment.NewTargetResource("SUB_ID", "RG_ID", "res", string(infra.AzureResourceTypeContainerApp)),
-			nil,
-			nil,
-			nil,
-			nil,
-			nil,
-		)
-
-		require.NoError(t, err)
-	})
-
-	t.Run("ValidateTypeLowerCaseSuccess", func(t *testing.T) {
-		_, err := NewContainerAppTarget(
-			nil,
-			nil,
-			environment.NewTargetResource(
-				"SUB_ID", "RG_ID", "res", strings.ToLower(string(infra.AzureResourceTypeContainerApp)),
+	tests := map[string]*serviceTargetValidationTest{
+		"ValidateTypeSuccess": {
+			targetResource: environment.NewTargetResource(
+				"SUB_ID",
+				"RG_ID",
+				"res",
+				string(infra.AzureResourceTypeContainerApp),
 			),
-			nil,
-			nil,
-			nil,
-			nil,
-			nil,
-		)
+			expectError: false,
+		},
+		"ValidateTypeLowerCaseSuccess": {
+			targetResource: environment.NewTargetResource(
+				"SUB_ID",
+				"RG_ID",
+				"res",
+				strings.ToLower(string(infra.AzureResourceTypeContainerApp)),
+			),
+			expectError: false,
+		},
+		"ValidateTypeFail": {
+			targetResource: environment.NewTargetResource("SUB_ID", "RG_ID", "res", "BadType"),
+			expectError:    true,
+		},
+	}
 
-		require.NoError(t, err)
-	})
+	for test, data := range tests {
+		t.Run(test, func(t *testing.T) {
+			mockContext := mocks.NewMockContext(context.Background())
+			serviceTarget := &containerAppTarget{}
+			serviceConfig := &ServiceConfig{}
 
-	t.Run("ValidateTypeFail", func(t *testing.T) {
-		_, err := NewContainerAppTarget(
-			nil,
-			nil,
-			environment.NewTargetResource("SUB_ID", "RG_ID", "res", "BadType"),
-			nil,
-			nil,
-			nil,
-			nil,
-			nil,
-		)
-
-		require.Error(t, err)
-	})
+			err := serviceTarget.validateTargetResource(*mockContext.Context, serviceConfig, data.targetResource)
+			if data.expectError {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
 }
 
 func Test_containerAppTarget_generateImageTag(t *testing.T) {
@@ -71,7 +66,15 @@ func Test_containerAppTarget_generateImageTag(t *testing.T) {
 	envName := "dev"
 	projectName := "my-app"
 	serviceName := "web"
+	serviceConfig := &ServiceConfig{
+		Name: serviceName,
+		Host: "containerapp",
+		Project: &ProjectConfig{
+			Name: projectName,
+		},
+	}
 	defaultImageName := fmt.Sprintf("%s/%s-%s", projectName, serviceName, envName)
+
 	tests := []struct {
 		name         string
 		dockerConfig DockerProjectOptions
@@ -79,7 +82,7 @@ func Test_containerAppTarget_generateImageTag(t *testing.T) {
 	}{
 		{"Default",
 			DockerProjectOptions{},
-			fmt.Sprintf("%s:azdev-deploy-%d", defaultImageName, mockClock.Now().Unix())},
+			fmt.Sprintf("%s:azd-deploy-%d", defaultImageName, mockClock.Now().Unix())},
 		{"ImageTagSpecified",
 			DockerProjectOptions{
 				Tag: NewExpandableString("contoso/contoso-image:latest"),
@@ -89,17 +92,12 @@ func Test_containerAppTarget_generateImageTag(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			containerAppTarget := &containerAppTarget{
-				env: environment.EphemeralWithValues(envName, map[string]string{}),
-				config: &ServiceConfig{
-					Name: serviceName,
-					Host: "containerapp",
-					Project: &ProjectConfig{
-						Name: projectName,
-					},
-					Docker: tt.dockerConfig,
-				},
-				clock: mockClock}
-			tag, err := containerAppTarget.generateImageTag()
+				env:   environment.EphemeralWithValues(envName, map[string]string{}),
+				clock: mockClock,
+			}
+			serviceConfig.Docker = tt.dockerConfig
+
+			tag, err := containerAppTarget.generateImageTag(serviceConfig)
 			assert.NoError(t, err)
 			assert.Equal(t, tt.want, tag)
 		})