+++ cli/azd/pkg/project/service_target_staticwebapp_test.go
@@ -4,52 +4,56 @@
 package project
 
 import (
+	"context"
 	"strings"
 	"testing"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	"github.com/azure/azure-dev/cli/azd/pkg/infra"
+	"github.com/azure/azure-dev/cli/azd/test/mocks"
 	"github.com/stretchr/testify/require"
 )
 
 func TestNewStaticWebAppTargetTypeValidation(t *testing.T) {
 	t.Parallel()
 
-	t.Run("ValidateTypeSuccess", func(t *testing.T) {
-		_, err := NewStaticWebAppTarget(
-			nil,
-			nil,
-			environment.NewTargetResource("SUB_ID", "RG_ID", "res", string(infra.AzureResourceTypeStaticWebSite)),
-			nil,
-			nil,
-		)
-
-		require.NoError(t, err)
-	})
-
-	t.Run("ValidateTypeLowerCaseSuccess", func(t *testing.T) {
-		_, err := NewStaticWebAppTarget(
-			nil,
-			nil,
-			environment.NewTargetResource(
-				"SUB_ID", "RG_ID", "res", strings.ToLower(string(infra.AzureResourceTypeStaticWebSite)),
+	tests := map[string]*serviceTargetValidationTest{
+		"ValidateTypeSuccess": {
+			targetResource: environment.NewTargetResource(
+				"SUB_ID",
+				"RG_ID",
+				"res",
+				string(infra.AzureResourceTypeStaticWebSite),
 			),
-			nil,
-			nil,
-		)
+			expectError: false,
+		},
+		"ValidateTypeLowerCaseSuccess": {
+			targetResource: environment.NewTargetResource(
+				"SUB_ID",
+				"RG_ID",
+				"res",
+				strings.ToLower(string(infra.AzureResourceTypeStaticWebSite)),
+			),
+			expectError: false,
+		},
+		"ValidateTypeFail": {
+			targetResource: environment.NewTargetResource("SUB_ID", "RG_ID", "res", "BadType"),
+			expectError:    true,
+		},
+	}
 
-		require.NoError(t, err)
-	})
+	for test, data := range tests {
+		t.Run(test, func(t *testing.T) {
+			mockContext := mocks.NewMockContext(context.Background())
+			serviceTarget := &staticWebAppTarget{}
+			serviceConfig := &ServiceConfig{}
 
-	t.Run("ValidateTypeFail", func(t *testing.T) {
-		_, err := NewStaticWebAppTarget(
-			nil,
-			nil,
-			environment.NewTargetResource("SUB_ID", "RG_ID", "res", "BadType"),
-			nil,
-			nil,
-		)
-
-		require.Error(t, err)
-	})
+			err := serviceTarget.validateTargetResource(*mockContext.Context, serviceConfig, data.targetResource)
+			if data.expectError {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
+		})
+	}
 }