+++ ext/vscode/src/utils/azureDevCli.ts
@@ -11,19 +11,19 @@ import { AsyncLazy } from './lazy';
 import { localize } from "../localize";
 import { AzExtErrorButton, IActionContext } from '@microsoft/vscode-azext-utils';
 import { isWindows } from './osUtils';
+import { setVsCodeContext } from './setVsCodeContext';
 
 // Twenty seconds: generous, but not infinite
 export const DefaultAzCliInvocationTimeout: number = 20 * 1000;
-const AzdInstallationUrl: string = 'https://aka.ms/azd-install';
-const AzdVersionCacheLifetime = 15 * 60 * 1000; // 15 minutes
+const AzdLoginCheckCacheLifetime = 15 * 60 * 1000; // 15 minutes
 
-enum AzdVersionCheckFailure {
-    NotInstalled = 1,
-    CannotDetermineVersion = 2
-}
-let userWarnedAzdMissing: boolean = false;
 let azdInstallAttempted: boolean = false;
-const azdVersionChecker = new AsyncLazy<string | AzdVersionCheckFailure>(getAzdVersion, AzdVersionCacheLifetime);
+const azdLoginChecker = new AsyncLazy<LoginStatus | undefined>(getAzdLoginStatus, AzdLoginCheckCacheLifetime);
+
+interface LoginStatus {
+    status: 'success' | 'unauthenticated' | string;
+    expiresOn?: string;
+}
 
 export type Environment = { [key: string]: string };
 export type AzureDevCli = {
@@ -33,39 +33,58 @@ export type AzureDevCli = {
 };
 
 export async function createAzureDevCli(context: IActionContext): Promise<AzureDevCli> {
-    const azdVersion = await azdVersionChecker.getValue();
-    if (azdVersion === AzdVersionCheckFailure.NotInstalled) {
+    const loginStatus = await azdLoginChecker.getValue();
+    if (!loginStatus) {
         context.errorHandling.suppressReportIssue = true;
         context.errorHandling.buttons = azdNotInstalledUserChoices();
-        userWarnedAzdMissing = true;
         throw new Error(azdNotInstalledMsg());
-    }
-    if (typeof azdVersion === 'string') {
-        context.telemetry.properties.azdVersion = azdVersion;
     }
 
     return createCli();
 }
 
-export function scheduleAzdInstalledCheck(): void {
-    const fiveSeconds = 5 * 1000;
+export function scheduleAzdSignInCheck(): void {
+    const oneSecond = 1 * 1000;
 
     setTimeout(async () => {
-        const ver = await azdVersionChecker.getValue();
+        const result = await azdLoginChecker.getValue();
 
-        if (ver === AzdVersionCheckFailure.NotInstalled && !userWarnedAzdMissing && !azdInstallAttempted) {
-            userWarnedAzdMissing = true;
-            const response = await vscode.window.showWarningMessage(azdNotInstalledMsg(), {}, ...azdNotInstalledUserChoices());
-            await response?.callback();
+        if (result !== undefined) {
+            // If we've reached this point, AZD is installed. We can set the VSCode context that the walk-through uses
+            await setVsCodeContext('hideAzdInstallStep', true);
+
+            // If the user is logged in, we can also set the login context
+            if (result.status === 'success') {
+                await setVsCodeContext('hideAzdLoginStep', true);
+            }
         }
-    }, fiveSeconds);
+    }, oneSecond);
+}
+
+export function scheduleAzdYamlCheck(): void {
+    const oneSecond = 1 * 1000;
+
+    setTimeout(async () => {
+        // Look for at most one file named azure.yml or azure.yaml, only at the root, to avoid perf issues
+        // If one exists, the scaffold step will be hidden from the walkthrough
+        const fileResults = await vscode.workspace.findFiles('azure.{yml,yaml}', undefined, 1);
+
+        if (fileResults?.length) {
+            await setVsCodeContext('hideAzdScaffoldStep', true);
+        }
+    }, oneSecond);
 }
 
 export function onAzdInstallAttempted(): void {
     azdInstallAttempted = true;
 
-    // Clear the install state so we'll check again at the next command
-    azdVersionChecker.clear();
+    // Clear the install+login state so we'll check again at the next command
+    azdLoginChecker.clear();
+}
+
+export function onAzdLoginAttempted(): void {
+    // Clear the install+login state so we'll check again at the next command
+    azdLoginChecker.clear();
 }
 
 function createCli(): AzureDevCli {
@@ -123,21 +142,17 @@ function getAzDevInvocation(): string[] {
     }
 }
 
-async function getAzdVersion(): Promise<string | AzdVersionCheckFailure> {
+async function getAzdLoginStatus(): Promise<LoginStatus | undefined> {
     const cli = createCli();
-    const command = cli.commandBuilder.withArgs(['version', '--output', 'json']).build();
-    let stdout: string;
+    const command = cli.commandBuilder.withArgs(['login', '--check-status', '--output', 'json']).build();
     try {
-        stdout = (await execAsync(command, cli.spawnOptions())).stdout;
-    } catch {
-        return AzdVersionCheckFailure.NotInstalled;
-    }
+        const stdout = (await execAsync(command, cli.spawnOptions())).stdout;
+        const result = JSON.parse(stdout) as LoginStatus;
 
-    try {
-        const versionSpec = JSON.parse(stdout) as { azd: { version: string } };
-        return versionSpec?.azd?.version || AzdVersionCheckFailure.CannotDetermineVersion;
+        return result;
     } catch {
-        return AzdVersionCheckFailure.CannotDetermineVersion;
+        // If AZD is not installed, return `undefined`
+        return undefined;
     }
 }
 
@@ -154,26 +169,20 @@ function normalize(env: NodeJS.ProcessEnv): Environment {
 }
 
 function azdNotInstalledMsg(): string {
-    return localize("azure-dev.utils.azd.notInstalled", "Azure Developer CLI is not installed. Would you like to install it?.");
+    return localize("azure-dev.utils.azd.notInstalled", "Azure Developer CLI is not installed. Would you like to install it? [Learn More](https://aka.ms/azd-install)");
 }
 
 function azdNotInstalledUserChoices(): AzExtErrorButton[] {
     const choices: AzExtErrorButton[] = [
         {
-            "title": localize("azure-dev.utils.azd.installNow", "Install"),
-            "callback": async () => {
+            title: localize("azure-dev.utils.azd.installNow", "Install"),
+            callback: async () => {
                 await vscode.commands.executeCommand("azure-dev.commands.cli.install", /* shouldPrompt: */ false);
             }
         },
         {
-            "title": localize("azure-dev.utils.azd.goToInstallUrl", "Learn More"),
-            "callback": async () => {
-                await vscode.env.openExternal(vscode.Uri.parse(AzdInstallationUrl));
-            }
-        },
-        {
-            "title": localize("azure-dev.utils.azd.later", "Later"),
-            "callback": () => { return Promise.resolve(); /* no-op */ }
+            title: localize("azure-dev.utils.azd.later", "Later"),
+            callback: () => { return Promise.resolve(); /* no-op */ }
         }
     ];
     return choices;