+++ templates/todo/api/js/src/config/index.ts
@@ -1,7 +1,6 @@
 import { AppConfig, DatabaseConfig, ObservabilityConfig } from "./appConfig";
 import dotenv from "dotenv";
-import { ChainedTokenCredential, DefaultAzureCredential } from "@azure/identity";
-import { AzureDeveloperCliCredential } from "../azureDeveloperCliCredential";
+import { DefaultAzureCredential } from "@azure/identity";
 import { SecretClient } from "@azure/keyvault-secrets";
 import { logger } from "../config/observability";
 import { IConfig } from "config";
@@ -53,7 +52,7 @@ const populateEnvironmentFromKeyVault = async () => {
 
     try {
         logger.info("Populating environment from Azure KeyVault...");
-        const credential = new ChainedTokenCredential(new AzureDeveloperCliCredential(), new DefaultAzureCredential());
+        const credential = new DefaultAzureCredential({});
         const secretClient = new SecretClient(keyVaultEndpoint, credential);
 
         for await (const secretProperties of secretClient.listPropertiesOfSecrets()) {